feat: Implement trash file management with undo and automatic expiration cleanup

+++ file-service/src/repositories/FileRepository.ts
@@ -1,8 +1,9 @@
 import File, { FilesAttributes } from "../db/models/File";
-import { Op } from "sequelize";
+import { Op, where } from "sequelize";
 import axios from "axios";
 import Category from "../db/models/Category";
 import { sendToQueue } from "../services/producer";
+import { deleteObject } from "../config/s3";
 
 class UploadRespository {
   upload = async (data: FilesAttributes) => {
@@ -47,7 +48,7 @@ class UploadRespository {
     sortBy: string,
     sortOrder: string
   ) => {
-    const whereClause: any = {
+    let whereClause: any = {
       userId,
     };
 
@@ -77,7 +78,10 @@ class UploadRespository {
 
     let data = await File.findAll({
       raw: true,
-      where: whereClause,
+      where: {
+        deletedAt: null,
+        ...whereClause
+      },
       limit: 10,
       offset,
       order: [[sortBy as string, (sortOrder || 'asc') as string]],
@@ -134,7 +138,7 @@ class UploadRespository {
     };
   };
 
-  deleteFile = async (fileId: string, token: string, offset: number) => {
+  deleteFile = async (fileId: string, token: string, offset: number, type: string) => {
     const deleteStarred = await axios.delete(`http://localhost:8082/api/file/starred/${fileId}?offset=${offset}`, {
       headers: {
         'Content-Type': 'application/json',
@@ -145,7 +149,7 @@ class UploadRespository {
       throw new Error("Starred service is unreachable");
     })
 
-    if(deleteStarred.status !== 200) {
+    if (deleteStarred.status !== 200) {
       throw new Error("Failed to delete starred-service")
     }
 
@@ -166,11 +170,24 @@ class UploadRespository {
       },
     }));
 
-    return await File.destroy({
-      where: {
-        id: fileId,
-      },
-    });
+    if (type === 'delete') {
+      return await File.destroy({
+        where: {
+          id: fileId,
+        },
+      });
+    } else {
+      const oneMonthFromNow = new Date();
+      oneMonthFromNow.setMonth(oneMonthFromNow.getMonth() + 1);
+
+      return await File.update({
+        deletedAt: oneMonthFromNow,
+      }, {
+        where: {
+          id: fileId,
+        },
+      });
+    }
   };
 
   totalFileSize = async (userId: string) => {
@@ -191,6 +208,63 @@ class UploadRespository {
     })
 
     return categories
+  }
+
+  getTrashFile = async (userId: string) => {
+    const trashFiles = await File.findAll({
+      where: {
+        userId,
+        deletedAt: {
+          [Op.ne]: null
+        }
+      }
+    })
+
+    return trashFiles
+  }
+
+  undoTrashFile = async (fileId: string) => {
+    const undoTrash = await File.update(
+      {
+        deletedAt: null,
+      },
+      {
+        where: {
+          id: fileId,
+        },
+      }
+    );
+
+    return undoTrash;
+  }
+
+  deleteExpiredFiles = async () => {
+    const thirsyDaysAgo = new Date();
+    thirsyDaysAgo.setDate(thirsyDaysAgo.getDate() - 30);
+
+    const expiredFiles = await File.findAll({
+      where: {
+        deletedAt: {
+          [Op.lt]: thirsyDaysAgo,
+        }
+      }
+    })
+
+    if(expiredFiles.length === 0) {
+      console.log("✅ No expired files found");
+      return; 
+    }
+
+    for (const file of expiredFiles) {
+      await deleteObject(file.location);
+      await file.destroy({
+        where: {
+          id: file.id
+        }
+      });
+    }
+
+    console.log(`🗑 Deleted ${expiredFiles.length} expired files`);
   }
 }
